Add a cache to boost ResultSet.findColumn performance
I have a use case where I have an intensive use of findColumn method.

When switching to sqlite-jdbc, I noticed really slower performance in my application.

Profiling the code, most of the time was spent in ResultSet.findColumn, so I applied the following modifications:
 - Add a simple map as a cache as in other db drivers (MySQL, PostgreSQL).
 - Remove the table prefix search in column name added in commit https://github.com/crawshaw/sqlitejdbc/commit/79da51f689ad94e7bf886bd27f8e83c3ce36d55d
 because we couldn't find a single test case where that was happening.

There is also an API change occurring with this commit. No exception is thrown
when columns have the same name.
This was not respecting the jdbc specification (java.sql.ResultSet) which states:

    When a getter method is called  with
    a column name and several columns have the same name,
    the value of the first matching column will be returned.