Warning message when readelf is not installed.
On an armhf system with no binutils installed and using a Java prior to version 8, armhf will not be detected and the wrong .so will be unpacked. Unfortunately this results in a crash with a FileNotFoundException, which is confusing (because the file is there and it has the right permissions).
This patch attempts to provide the user with a more meaningful message.