Handle old SQLite releases that don't follow the current naming convention.
Fetch SQLite sources even when the archive is named after the dotted-version
number (as opposed to the "encoded" version number).

Also cope with SQLite archives that don't extract into a version-encoded
subdirectory by extracting into a temporary sub-directory and renaming it
so it complies with the current naming scheme.

This allows build back to 3.6.11. However, tests fail (because they assume
functionality that was no-existent at that time)