Avoid any hardcoded file permissions
On Unix systems hardcoded file permissions should be avoided: When
creating files the umask of the current user should apply. Hardcoding
permissions prevents SQLite files to be created group-writable for
instance (umask = 0002)

To achieve this, configure the default SQLite permissions to 0666. This
grants read & write to all, then the umask applies, e.g. 0002, resulting
in permissions 0664.

Note that I'm not sure how that affect Windows & Mac. I think for
Windows it should be ok as it's using a different permission scheme, and
for Mac it should behave the same as Unix .