Skip synchronized block in close() if ResultSet is already closed (#358)
Fixes #349

From profiling our app in production we noticed that our finalizer thread was periodically getting the synchronized lock on the DB (introduced in #215).  By the time our ResultSet hits the finalizer queue the ResultSet has already been closed, so there's no need to acquire the lock to reset the statement pointer.  As part of this I've removed what I believe to be dead code around checking for the nullability of the stmt member on the ResultSet.