Support temporary databases with no file name (#396)
When no database is specified, SQLite is supposed to create a temporary
database, i.e. residing in memory but able to be flushed to disk under
memory pressure.
Actually, the driver forces SQLite to open an in-memory database
instead. These databases cannot be flushed to disk and might come out of
memory.