Add suport for external archive amalgamation (#464)
By passing SQLITE_OBJ=/path/to/libsqlite3.a and SQLITE_HEADER=/path/to/sqlite3.h
one can circumvent SQLite to be downloaded and custom compiled.

Users can have SQLite globally installed, reusing that for JDBC driver instead
of having a version deviating from the rest of the system.

This closes #464