feat(native-image): resource optimization and configuration to export native lib
This replaces the resource-config.json and jni-config.json files by a org.graalvm.nativeimage.hosted.Feature implementation.
The registered JNI configuration is unchanged, but is now configured using the native-image java API. This is easier to maintain than the json files.
The resource config is different, as previously all library resources were included. Now only the library relevant for the target runtime of the native-image compilation is included as a resource. To do this, the existing logic in SQLiteJDBCLoader was reused. Additionally the metadata resources containing version info and properties are read at build-time in static code blocks which avoid the need to include those resources and avoid the need to do the IO operations at runtime.
Finally, I added a new org.sqlite.lib.exportPath which takes in a directory that is used by the SqliteJdbcFeature to export the JNI library to at build-time. If this property is set, the library will no longer be included as a resource and it becomes the responsibility of the end user to provide the exported library next to the native-image executable.
The benefit of doing this is that the library no longer needs to be unpacked at runtime and again some IO operations can be eliminated.

Co-authored-by: Andrew Pikler <andrew.pikler@earnix.com>