fix(jdbc): reintroduce improved support for Statement#getGeneratedKeys
Adds back support for the getGeneratedKeys functionality previously removed in
commit 332edfb4dc3f4dcb463f0d1d8417f9c92519e609. The new implementation attaches the
generated keys directly to the Statement so that it is safe to use in an interleaved
manner and accross threads.

This does not address tracking inserts of multiple values in the same statement (#468)
but that functionality was never previously available. It could potentially be added in
future enhancements.

Resolves: #329, #963, #965, #980, #1018, #1028